python/pyproject.toml
[tool.poetry]
name = "adb-auto-player"
version = "0.0.0"
description = ""
authors = ["yulesxoxo <https://github.com/yulesxoxo/yulesxoxo>"]
readme = "README.md"
package-mode = true

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
pillow = "^11.0.0"
adbutils = "^2.8.0"
opencv-python = "^4.10.0.84"
pydantic = "^2.10.5"
av = "^14.1.0"

[tool.poetry.group.dev.dependencies]
nuitka = "^2.5.9"
pytest = "^8.3.4"
commitizen = "^4.2.2"

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py312"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
select = [
    "F",   # pyflakes
    "D",   # pydocstyle
    "E",   # pycodestyle error
    "N",   # pep8-naming
    "RUF", # ruff
    "UP",  # pyupgrade
    "PL",  # pylint
]
ignore = [
    "D203",  # incorrect-blank-line-before-class
    "D406",  # missing-new-line-after-section-name
    "D407",  # missing-dashed-underline-after-section
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
adb-auto-player = "adb_auto_player.main:main"
